pyproject.toml
[project]
name = "labs-day6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.2",
]
